// src/shared/validateField/validate-field.enum.ts
/* eslint-disable prettier/prettier */
export enum EnumValidateField {
    INVALID_EXIST_EMAIL_USERNAME_ADMIN = 'INVALID_EXIST_EMAIL_USERNAME_ADMIN',
    INVALID_EXIST_EMAIL_USER = 'INVALID_EXIST_EMAIL_USER',
    INVALID_EXIST_PHONE_USER = 'INVALID_EXIST_PHONE_USER',
    INVALID_CURRENT_PASSWORD = 'INVALID_CURRENT_PASSWORD',
    INVALID_CURRENT_CONFIRM_PASSWORD = 'INVALID_CURRENT_CONFIRM_PASSWORD',
    NOT_EXIST_PROVINCE = 'NOT_EXIST_PROVINCE',
    NOT_EXIST_DISTRICT = 'NOT_EXIST_DISTRICT',
    NOT_EXIST_WARD = 'NOT_EXIST_WARD',
    INVALID_NOT_MATCH_PASSWORD = 'INVALID_NOT_MATCH_PASSWORD',
    IS_NOT_EMPTY = 'IS_NOT_EMPTY',
    IS_STRING = 'IS_STRING',
    MAX_LENGTH = 'MAX_LENGTH',
    MIN_LENGTH = 'MIN_LENGTH',
}